backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
yt-dlp==2025.2.19
mutagen==1.47.0
python-multipart==0.0.20
aiofiles==24.1.0
Pillow==11.1.0
